% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MABM_report.R
\name{MABM_report}
\alias{MABM_report}
\title{MABM annual report creation.}
\usage{
MABM_report(station = NULL, year = lubridate::year(Sys.Date()),
  out_dir = ".", MABM_dir = NULL)
}
\arguments{
\item{station}{character string indicating the MABM station (e.g., refuge, ES office) for
which to produce a report.  Default (NULL) prompts the user to select from a list of all
available MABM stations}

\item{year}{integer indicating the year for which to generate the annual report; defaults
to the current year}

\item{out_dir}{character string indicating path to directory where final report *.pdf
should be deposited.  Default is the current working directory}

\item{MABM_dir}{character string indicating path to directory where *.xlsx output of
\code{\link{setup_MABM_reports}} is stored.  Currently, the macro implanted in the MABM
database only outputs to C:/temp, so this argument is not useful at the moment}
}
\description{
This function generates an annual report for one or more MABM stations for, by default,
 the current calendar year.  The report includes, for each route associated with a
 given station, (1) an overview of the year's efforts, (2) a night-by-night tabular
 summary with notes, if present, (3) a night-by-night tabular summary of species
 detections, (4) a tabular and graphical comparison of species detections from the
 current season with previous season(s), and (5) a map of the route indicating the
 locations of all bat detections for the season.
}

