% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MABMreportr.R
\docType{package}
\name{MABMreportr}
\alias{MABMreportr}
\alias{MABMreportr-package}
\title{\code{MABMreportr} package}
\description{
See the \href{https://github.com/adamdsmith/MABMreportr}{README} on GitHub
}

