% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_MABM_reports.R
\name{setup_MABM_reports}
\alias{setup_MABM_reports}
\title{Export data from MABM database necessary for report preparation}
\usage{
setup_MABM_reports(access_exe = NULL, MABM_access_db = NULL,
  export_dir = NULL)
}
\arguments{
\item{access_exe}{character string indicating the path to the MS Access executable.  Default (NULL)
attempts to find it in a couple of common places and, if unsuccessful, prompts the user to browse
to the executable.}

\item{MABM_access_db}{character string indicating the path to the MABM MS Access database.
Default (NULL) prompts the user to browse to the file.}

\item{export_dir}{character string indicating desired location to output *.xlsx file from
Access database; currently not used}
}
\value{
None.This function calls a macro within the MABM database to export
 MS Excel spreadsheets to a specified directory for use in annual report creation.
}
\description{
This function calls macros inside the MABM database to export data necessary for the
creating of annual reports for MABM routes.  It only needs to be run a single time after
the data for a given survey year has been entered into the database.
}

